// mobile-app/src/types/Property.ts
export interface Property {
  _id: string;
  title: string;
  slug: string;
  price: string;
  priceFrom?: string;
  priceTo?: string;
  propertyType: 'buy' | 'rent' | 'commercial' | 'plot' | 'pg' | 'new' | string;
  subType: string;
  category?: string;
  street1?: string;
  street2?: string;
  street3?: string;
  street4?: string;
  area: string;
  city: string;
  state: string;
  district?: string;
  pincode?: string;
  latitude?: string;
  longitude?: string;
  facing?: string;
  beds?: string[];
  baths?: string;
  sqft: string;
  description: string;
  badge?: string;
  isVerified?: boolean;
  isSold?: boolean;
  isNewProject?: boolean;
  foodAvailable?: boolean;
  acAvailable?: boolean;
  amenities: string[];
  images: string[];
  contactPhone?: string;
  whatsappNumber?: string;
  landmarks?: { name: string; distance: string }[];
  brochureUrl?: string;
  createdAt: string;
}